Week4_lesson/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
groq==0.4.1
gtts==2.4.0
python-dotenv==1.0.0
aiofiles==23.2.1
